+++ src/icemac/ab/calendar/browser/interfaces.py
@@ -5,12 +5,10 @@ import zope.schema
 
 
 class ICalendarLayer(z3c.layer.pagelet.IPageletBrowserLayer):
-
     """Calendar browser layer."""
 
 
 class IDatetime(zope.interface.Interface):
-
     """Object interface to edit datetime data."""
 
     whole_day_event = zope.schema.Bool(
@@ -28,7 +26,43 @@ class IDatetime(zope.interface.Interface):
 
 
 class IEventDatetime(zope.interface.Interface):
-
     """Interface to edit event's date and time."""
 
     datetime = zope.schema.Object(title=_('datetime'), schema=IDatetime)
+
+
+class IEventDescription(zope.interface.Interface):
+    """Description of a single event which can be rendered."""
+
+    context = zope.interface.Attribute('IEvent this description is based on.')
+
+    kind = zope.interface.Attribute('event kind')
+    datetime = zope.interface.Attribute('datetime.datetime object')
+    prio = zope.interface.Attribute(
+        'Event descriptions for the same `datetime` and `kind` with a higher '
+        '`prio` override the ones with lower `prio`.')
+
+    whole_day = zope.interface.Attribute(
+        'Event is the whole day, so do not display time.')
+
+    persons = zope.interface.Attribute(
+        'Komma separated list of person names belonging to the event.')
+
+    def getText(lang=None):
+        """Textual description of the event.
+
+        If `lang` is not `None` a hyphenation dict for this language is
+        looked up. This might raise a `LookupError`. Otherwise the text is
+        hyphenated for HTML.
+        """
+
+    def getInfo(lang=None):
+        """List of additional information about the event.
+
+        The contents of the list are defined in master data of calendar.
+
+        """
+
+
+class UnknownLanguageError(LookupError):
+    """Error indicating an unknown laguage."""